requirements.txt
requests>=2.31.0
pyyaml>=6.0
edge-tts>=6.1.0
